refactor(zustand->RQ): moved out remaining methods and constants out of the zustand store.
I moved out the remaining methods and constants that I was managing with zustand because it doesn't make sense anymore since I'm managing server state with react query.
