+++ src/constants.js
@@ -1,0 +1,1 @@
+export const habitMilestones = [7, 14, 21, 30, 60, 120, 365];
