+++ src/utils.ts
@@ -1,5 +1,6 @@
 import clsx, { ClassValue } from "clsx";
 import { twMerge } from "tailwind-merge";
+import { habitMilestones } from "./constants";
 
 const MONTHS = [
   "January",
@@ -159,3 +160,104 @@ export const isToday = (dirtyDate: Date | string): boolean => {
 export function cn(...inputs: ClassValue[]) {
   return twMerge(clsx(inputs));
 }
+
+/**
+ * Returns the current streak for a habit, based on the provided months array.
+ * @param entries - An array of entries containing the id, date and state.
+ * @returns streak - The current streak for the habit.
+ */
+export const getHabitStreak = (entries: any[]) => {
+  const today = new Date();
+  let streak = 0;
+
+  const lastMonth = entries.at(-1).date;
+
+  const reversedEntries = entries
+    .flat()
+    .reverse()
+    .slice(daysInCurrentMonth() - today.getDate());
+
+  for (let entry of reversedEntries) {
+    if (entry.state !== "completed") {
+      break;
+    }
+    streak += 1;
+  }
+  return streak;
+};
+
+export const generatePendingHabitEntries = (datesArray) => {
+  return datesArray.map((date) => ({
+    date,
+    state: "pending",
+  }));
+};
+
+export const getNextMonthPendingHabitEntries = (prevDate) => {
+  const date = new Date(prevDate);
+  date.setMonth(date.getMonth() + 1);
+  const monthDates = getAllDaysInMonth(date.getFullYear(), date.getMonth());
+
+  return generatePendingHabitEntries(monthDates);
+};
+
+/**
+ *
+ * @param {Date} currentDate the current date object
+ * @param {Date} prevMonthAdded the last month added date object
+ * @returns {Boolean} whether it should or shouldn't add the next month
+ */
+export const shouldAddNextMonth = (currentDate, prevMonthAdded) => {
+  const isYearBeforeCurrent =
+    prevMonthAdded.getFullYear() < currentDate.getFullYear();
+  const isMonthBeforeCurrent =
+    prevMonthAdded.getFullYear() === currentDate.getFullYear() &&
+    prevMonthAdded.getMonth() < currentDate.getMonth();
+  return isYearBeforeCurrent || isMonthBeforeCurrent;
+};
+
+/**
+ * Check if the current streak reached a milestone/s and return the new milestones.
+ * @param {number} currentStreak - The current streak of the habit.
+ * @param {Array} badges - The badges of the habit.
+ * @returns {Array} - The new milestones.
+ * */
+export const checkForNewMilestones = (
+  currentStreak: number,
+  badges: number[]
+) => {
+  const newMilestones = habitMilestones.filter(
+    (milestone) => milestone <= currentStreak && !badges.includes(milestone)
+  );
+  return newMilestones.length > 0 ? newMilestones : null;
+};
+
+export const getPast7DaysEntries = (entries, currentDate) => {
+  const currentEntryIndex = entries.findIndex(
+    (entry) => startOfDay(entry.date).getTime() === currentDate.getTime()
+  );
+
+  let lastWeek = entries.slice(
+    currentEntryIndex + 1 - 7,
+    currentEntryIndex + 1
+  );
+
+  if (currentEntryIndex < 7) {
+    lastWeek = entries.slice(0, currentEntryIndex + 1);
+  }
+
+  if (lastWeek.length < 7) {
+    //if the current habit does not have data for the previous month,
+    //generate placeholder values.
+
+    //get prev 7 days based on the first available date
+    const previousPlaceholderDates = getPast7Days(new Date(lastWeek[0].date))
+      .map((date) => ({ id: date, date, state: "pending" }))
+      .sort((a, b) => a.id.getDate() - b.id.getDate())
+      .slice(lastWeek.length);
+
+    lastWeek = previousPlaceholderDates.concat(lastWeek);
+  }
+
+  return lastWeek;
+};
